Refined players shooting with new script
Made custom 'PlayerShooting' script thats basically the same as Dylan's 'PlayerShooter' script. But allows me to add on in different ways later on and add more organization. Also removed old 'PlayerMovement' script from player prefab so it only uses physics to move now. Also made a new 'PooBullet' script that will hopefully add a force in the direction the player is facing. Also For now it only uses inputs to add a force in that direction. So it won't shoot in the direction you are facing and only adds a force to the direcion you are moving. Disabled the 'SimpleForce2D' script from Dylan.