Made a health system for player
Changed lives to be game objects instead of canvas UI to be able to control them more. I worked on player health script. I made a switch case that referenced the number of lifes the player has. If it has a certain amount change the active player life game objects. I also made a collision detection that will check if the goose (player) collided with a enemy then change its position to far away. While also invoking the death cases in a bit of a delay.