Made start menu/scene functional & added in a simple duck sprite
Changed the duck sprite to a simple duck with no animations yet. Created a new scene for the tutorial but it is empty for now, you can load it and exit it for functional purposes. Added all the calls to scene manager in the main menu. Also changed the 'sound button' to 'tutorial'. Added all relevant scenes into build settings and all functions into a scene manager prefab.